Actually export formatAddress
It was exported from Utilities, but Utilities is not exported (which
makes sense)
